Removing old archaeic remenants of our previous stylesheet system
Yeah, most stylesheets shouldn't need a "stylesheet system";
unfortunately, this stylesheet from back in the day had a bunch of files
with it. Not a good idea. Just look at all the different parts that
worked together to make it work... and it wasn't even pretty.
This old stylesheet looked like something made by a toddler,
specifically a toddler who really enjoys watching Teletubbies.

So, it's officially becoming gone now; I don't believe anything in the
deep bowels of the Internet is still using it.  About time.